+++ src/main/java/racingcar/GameRun.java
@@ -1,6 +1,6 @@
 package racingcar;
 
-import racingcar.game.view.RacingStadium;
+import racingcar.view.RacingStadium;
 
 public class GameRun implements Runnable {
